Generaliyation of time-space interpolator

+++ bmcs_shear/dic_crack/cached_array.py
@@ -2,7 +2,8 @@
 from pathlib import Path
 import numpy as np
 
-def cached_array(source_name, names=[], data_dir_trait='data_dir', refresh=False):
+def cached_array(names=[], data_dir_trait='data_dir', refresh=False, 
+                 timestamp_name='beam_params.txt'):
     """Load the contents of an array with the specified name if it exists and is newer than source.
 
     Args:
@@ -15,7 +16,8 @@ def cached_array(source_name, names=[], data_dir_trait='data_dir', refresh=False
         function: Wrapper function that checks if force_array_refresh is True to ignore cached value.
 
     Examples:
-        @cached_array(source_name="beam_param_file", names=['tstring', 'time', 'F'], data_dir_trait='data_dir')
+        @cached_array(names=['tstring', 'time', 'F'], data_dir_trait='data_dir', 
+                      timestamp_name="beam_params.txt")
         def access_cached_array(self):
             # Function implementation
     """
@@ -27,12 +29,16 @@ def cached_array(source_name, names=[], data_dir_trait='data_dir', refresh=False
             if force_array_refresh:
                 print('forced refreshing of cached arrays for', self)
 
-            source_file = Path(getattr(self, source_name))
+            obj_name = getattr(self, "name")
             data_dir = getattr(self, data_dir_trait)
             cache_dir = Path(data_dir) / "cache"
+            timestamp_file = data_dir / timestamp_name
+            if timestamp_file.exists() == False:
+                raise FileNotFoundError(f"timestamp file {timestamp_file} not found")
+
             names_ = [names] if isinstance(names, str) else names
-            cached_attr_name = f'_cached_{"_".join(names_)}'
-            cache_file = cache_dir / f"cached_{'_'.join(names_)}.npz"
+            cached_attr_name = f'_{obj_name}_cached_{"_".join(names_)}'
+            cache_file = cache_dir / f"{cached_attr_name}.npz"
             if force_array_refresh and cache_file.exists():
                 cache_file.unlink()
             if hasattr(self, cached_attr_name):
@@ -40,8 +46,8 @@ def cached_array(source_name, names=[], data_dir_trait='data_dir', refresh=False
             if cache_dir.is_dir():
                 if cache_file.exists():
                     cache_ctime = cache_file.stat().st_ctime
-                    source_ctime = source_file.stat().st_ctime
-                    if cache_ctime > source_ctime:
+                    timestamp_ctime = timestamp_file.stat().st_ctime
+                    if cache_ctime > timestamp_ctime:
                         loaded = np.load(cache_file)
                         return loaded[names] if isinstance(names, str) else [loaded[name] for name in names_]
             else:
